+++ packages/applications/ssr/src/components/pages/réseau/raccordement/modifier/modifierPropositionTechniqueEtFinancière/ModifierPropositionTechniqueEtFinancière.form.tsx
@@ -1,0 +1,83 @@
+'use client';
+
+import { FC, useState } from 'react';
+import { useRouter } from 'next/navigation';
+import Button from '@codegouvfr/react-dsfr/Button';
+
+import { Routes } from '@potentiel-applications/routes';
+import { now } from '@potentiel-libraries/iso8601-datetime';
+
+import { Form } from '@/components/atoms/form/Form';
+import { InputDate } from '@/components/atoms/form/InputDate';
+import { UploadDocument } from '@/components/atoms/form/UploadDocument';
+import { SubmitButton } from '@/components/atoms/form/SubmitButton';
+
+import { modifierPropositionTechniqueEtFinancièreAction } from './modifierPropositionTechniqueEtFinancière.action';
+import { ModifierPropositionTechniqueEtFinancièrePageProps } from './ModifierPropositionTechniqueEtFinancière.page';
+
+type ModifierPropositionTechniqueEtFinancièreFormProps =
+  ModifierPropositionTechniqueEtFinancièrePageProps;
+
+export const ModifierPropositionTechniqueEtFinancièreForm: FC<
+  ModifierPropositionTechniqueEtFinancièreFormProps
+> = ({
+  identifiantProjet,
+  raccordement: {
+    reference,
+    propositionTechniqueEtFinancière: { dateSignature, propositionTechniqueEtFinancièreSignée },
+  },
+}) => {
+  const router = useRouter();
+  const [validationErrors, setValidationErrors] = useState<Array<string>>([]);
+
+  return (
+    <Form
+      method="POST"
+      encType="multipart/form-data"
+      action={modifierPropositionTechniqueEtFinancièreAction}
+      onSuccess={() => router.push(Routes.Raccordement.détail(identifiantProjet))}
+      heading="Modifier la proposition technique et financière"
+      onValidationError={(validationErrors) => setValidationErrors(validationErrors)}
+    >
+      <p className="my-2 p-0">Référence du dossier de raccordement : {reference}</p>
+
+      <input type="hidden" name="identifiantProjet" value={identifiantProjet} />
+      <input type="hidden" name="referenceDossierRaccordement" value={reference} />
+
+      <InputDate
+        id="dateSignature"
+        label="Date de signature"
+        nativeInputProps={{
+          type: 'date',
+          name: 'dateSignature',
+          max: now(),
+          defaultValue: dateSignature,
+          required: true,
+        }}
+      />
+
+      <UploadDocument
+        label="Proposition technique et financière signée"
+        name="propositionTechniqueEtFinanciereSignee"
+        required
+        state={
+          validationErrors.includes('propositionTechniqueEtFinanciereSignee') ? 'error' : 'default'
+        }
+        documentKey={propositionTechniqueEtFinancièreSignée}
+      />
+
+      <div className="flex flex-col md:flex-row gap-4 m-auto">
+        <Button
+          priority="secondary"
+          linkProps={{
+            href: Routes.Raccordement.détail(identifiantProjet),
+          }}
+          iconId="fr-icon-arrow-left-line"
+        >
+          Retour aux dossiers de raccordement
+        </Button>
+        <SubmitButton>Modifier</SubmitButton>
+      </div>
+    </Form>
+  );
+};
